@@ -9,7 +9,7 @@ namespace AMQSongProcessor.UI
 	public static class TaskbarProgress
 	{
 		private static readonly Lazy<ITaskbarList3> _TaskbarInstance
-			= new Lazy<ITaskbarList3>(() => (ITaskbarList3)new TaskbarInstance());
+			= new(() => (ITaskbarList3)new TaskbarInstance());
 		private static readonly bool _Windows7OrGreater =
 			FileUtils.IsWindows && Environment.OSVersion.Version >= new Version(6, 1);
 		private static TaskbarStates _TaskbarState;
